% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_dtest.R
\name{D.test}
\alias{D.test}
\title{Determination of significance between models}
\usage{
D.test(nsModel, sModel, sign = 0.05)
}
\arguments{
\item{nsModel}{non-stationary model (gpd object).}

\item{sModel}{stationary model (gpd object).}

\item{sign}{(optional) significance level (0.05 as default).}
}
\value{
D.test result.
}
\description{
D.test determines the significance of non-stationary over stationary POT model based on the likelihood test, at significant level sign.
}
\details{
Hypothesis: the NSPOT model is better -> yields TRUE; H0: the stationary model is better -> yields FALSE.
}
