% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_findevents.R
\name{findevents}
\alias{findevents}
\title{Find precipitation events}
\usage{
findevents(data, thresholdA = 0, thresholdB = 0)
}
\arguments{
\item{data}{(numeric) data series of daily precipitation values.}

\item{thresholdA}{minimum amount of precipitation considered as rainy day (default 0).}

\item{thresholdB}{same as thresholdA, used to set the ending of an event.}
}
\value{
matrix with indices corresponding to beginning, endings and length of the events.
}
\description{
This function identifies consecutive rainy days based on adily precipitation series.
}
