% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_mapq.R
\name{mapq}
\alias{mapq}
\title{Mapping univariate and multivariate NSPOT results}
\usage{
mapq(
  sts,
  ind_names,
  res,
  path,
  brks = c(1.25, 2, 3, 4),
  brks_time = c(1.1, 1.2, 1.3, 1.5),
  country = NULL
)
}
\arguments{
\item{sts}{A matrix or data frame containing (all) the stations information.
Must have, at least: ID, NAME, LON, LAT.}

\item{ind_names}{names of the indices.}

\item{res}{results object (list) from a call to the nspot() function.}

\item{path}{Directory where the results in pdf format will be saved.}

\item{brks}{Vector (numeric). Breaks in which indices values (not time) will
be categorized for representation purposes.}

\item{brks_time}{Vector (numeric). Breaks in which the temporal index values
(temporal trends) will be categorized for representation purposes.}

\item{country}{uses administrative boundaries of provinces (or equivalent) of
a specific country. If NULL, only country boundaries are shown.}
}
\value{
collection of PDF files representing the univariate and multivariate
trends, by indices.
}
\description{
This function creates maps of univariate and multivariate
differences between q100 positive and negative values of the indices,
based on results from a call to the nspot() function.
}
