% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_nspot.R
\name{nspot}
\alias{nspot}
\title{Non-Stationary Peak-over-threshold analysis (NSPOT)}
\usage{
nspot(
  pcp,
  ind,
  sts,
  dates,
  tstep = "annual",
  thres = 0.9,
  var = "intensity",
  path = NULL,
  isEvent = TRUE,
  method = "gpd",
  ind_names
)
}
\arguments{
\item{pcp}{Vector containing the ID of the station (first element) and the daily precipitation data series (rest of the elements).}

\item{ind}{Matrix or data.frame with daily indices, by columns.}

\item{sts}{Matrix or data.frame containing (all) the stations information. Must have, at least: ID, NAME, LON, LAT.}

\item{tstep}{Time step considered for the analysis. It can be a character ("annual", default) or a numeric vector containing the months.}

\item{thres}{numeric vector indicating the percentile (from 0 to 1) over which events will be considered for the analysis.}

\item{var}{Variable to analyze computed from the extreme precipitation events. Possible values are: "magnitude" (total precipitation of the event); "intensity" (mean daily precipitation of the event); "maximum_intensity" (maximum daily value of the event) and "duration" (number of days of the event).}

\item{path}{Directory where results in pdf will be saved. If NULL, no files will be created.}

\item{isEvent}{Logical. If TRUE, precipitation events (consecutive rainy days) will be considered instead of single rainy days (FALSE).}

\item{method}{Method used for fitting variables. At this moment, only "gpd" (Maximum-likelihood Fitting for the GPD Model) is available.}

\item{ind_names}{Character vector with the names of the indices to consider in the analysis. The names must match with one or more column names in "ind".}
}
\value{
A list with resulting parameters is returned. If path is defined, a pdf file is created.
}
\description{
This function analyzes the individual and combined influence of climatic indices on extreme precipitation events.
}
\examples{


}
